improve project integration test

+++ crates/app_core/src/dto/projects_dto.rs
@@ -2,17 +2,17 @@ use super::project_with_thumbnail_dto::ProjectWithThumbnailDto;
 
 #[derive(Default, Debug, Clone)]
 pub struct ProjectsDto {
-    pub page: i32,
-    pub size: i32,
-    pub total_pages: i32,
+    pub page: i64,
+    pub size: i64,
+    pub total_pages: i64,
     pub projects: Vec<ProjectWithThumbnailDto>,
 }
 
 impl ProjectsDto {
     pub fn new(
-        page: i32,
-        size: i32,
-        total_pages: i32,
+        page: i64,
+        size: i64,
+        total_pages: i64,
         projects: Vec<ProjectWithThumbnailDto>,
     ) -> Self {
         Self {
